+++ src/components/Space/Profile.module.scss
@@ -116,6 +116,12 @@
   }
 }
 
+.kebobBtn {
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+}
+
 @media screen and (max-width: 768px) {
   .container {
     box-shadow: none;
